Initial commit
The programs ordenar.py and calificar.py work now. The auxiliary files
LoggingFileHandlers.py and levenshtein.py are there to complement the
module. The configuration file logging.conf is there to automatically
manage logging. The file config.ini.example is there to change how the
program behaves in different environments.